pd: split genesis code into modules
