+++ pd/src/consensus/worker.rs
@@ -1,12 +1,10 @@
-use std::collections::BTreeSet;
-
 use anyhow::{anyhow, Result};
 use futures::StreamExt;
 use metrics::absolute_counter;
-use penumbra_crypto::{asset, merkle::NoteCommitmentTree, note};
+use penumbra_crypto::{asset, merkle::NoteCommitmentTree};
 use penumbra_proto::Protobuf;
 use penumbra_stake::{
-    IdentityKey, ValidatorState, ValidatorStatus, STAKING_TOKEN_ASSET_ID, STAKING_TOKEN_DENOM,
+    ValidatorState, ValidatorStatus, STAKING_TOKEN_ASSET_ID, STAKING_TOKEN_DENOM,
 };
 use penumbra_transaction::Transaction;
 use tendermint::abci::{self, ConsensusRequest as Request, ConsensusResponse as Response};
